refactor(calendar): add data and presentation folders
