+++ lib/features/calendar/data/repository/calendar_repository.dart
@@ -2,11 +2,11 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "../../../../api_base_rest/client/json.dart";
-import "../../../../api_base_rest/translations/translate.dart";
-import "../../../../config/env.dart";
+import "../../../../../api_base_rest/client/json.dart";
+import "../../../../../api_base_rest/translations/translate.dart";
+import "../../../../../config/env.dart";
 
-import "../calendar_view.dart";
+import "../../presentation/calendar_view.dart";
 import "../model/calendar_data.dart";
 part "calendar_repository.g.dart";
